requirements.txt
accelerate==0.28.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
coverage==7.4.3
cupy-cuda12x==12.1.0
DataProperty==1.0.1
datasets==2.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
evaluate==0.4.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastrlock==0.8.2
filelock==3.13.1
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
identify==2.5.35
idna==3.6
interegular==0.3.3
Jinja2==3.1.3
joblib==1.3.2
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lark==1.1.9
llvmlite==0.42.0
lxml==5.1.0
MarkupSafe==2.1.5
mbstrdecoder==1.1.3
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
mypy==1.9.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
nodeenv==1.8.0
numba==0.59.1
numexpr==2.9.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
openai==1.14.3
optimum==1.17.1
outlines==0.0.34
packaging==24.0
pandas==2.2.1
pathvalidate==3.2.0
pillow==10.2.0
platformdirs==4.2.0
pre-commit==3.6.2
prometheus_client==0.20.0
protobuf==5.26.0
psutil==5.9.8
pyarrow==15.0.1
pyarrow-hotfix==0.6
pydantic==2.6.4
pydantic_core==2.16.3
pynvml==11.5.0
pytablewriter==1.2.0
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
ray==2.9.3
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
responses==0.18.0
rouge-score==0.1.2
rpds-py==0.18.0
s3cmd==2.4.0
safetensors==0.4.2
scipy==1.12.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
sqlitedict==2.1.0
starlette==0.36.3
sympy==1.12
tabledata==1.3.3
tcolorpy==0.1.4
termcolor==2.4.0
tokenizers==0.15.2
torch==2.1.2
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.2
tqdm-multiprocess==0.0.11
transformers==4.39.1
triton==2.1.0
typepy==1.3.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.1
vllm==0.3.3
watchfiles==0.21.0
websockets==12.0
xformers==0.0.23.post1
xxhash==3.4.1
yarl==1.9.4
zstandard==0.22.0